+++ Doom/src/Doom/Render/Mesh.h
@@ -8,31 +8,30 @@
 struct DOOM_API Mesh {
 public:
 	std::string name;
-	double* verteces = nullptr;
+	float* verteces = nullptr;
 	uint32_t vertecesSize = 0;
-	double* vertecesForNormals = nullptr;
+	float* vertecesForNormals = nullptr;
 	uint32_t vertecesSizeForNormals = 0;
 	uint32_t* indicesForNormals = nullptr;
 	uint32_t indicesSizeForNormals = 0;
 	uint32_t* indices = nullptr;
 	uint32_t indicesSize = 0;
-	double* normals = nullptr;
+	float* normals = nullptr;
 	uint32_t normalsSize = 0;
 	uint32_t meshSize = 0;
-	double* mesh = nullptr;
+	float* mesh = nullptr;
 	uint32_t uvSize = 0;
-	double* uv;
+	float* uv;
 	uint32_t uvSizeForVert = 0;
-	double* uvForVert;
+	float* uvForVert;
 	uint32_t uvIndexSize = 0;
 	uint32_t* uvIndex;
-	double* btangent;
-	double* tangent;
+	float* btangent;
+	float* tangent;
 
 	Mesh(std::string name) : name(name){}
 
 	~Mesh() {
-		//std::cout << "Mesh " << name.c_str() << " destroyed\n";
 		delete[] mesh;
 	}
 };
